mysite/requirements.txt
anyascii==0.3.1
asgiref==3.5.2
beautifulsoup4==4.9.3
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.0.8
django-filter==21.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
l18n==2021.3
openpyxl==3.0.10
Pillow==9.3.0
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.2.1
telepath==0.3
urllib3==1.26.12
wagtail==3.0.3
wagtail-nav-menus==3.9.0
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
